Split ImageGalleryController into ImageController and ImageGalleryController
