+++ src/AppBundle/Controller/ImageGalleryController.php
@@ -3,10 +3,7 @@
 namespace AppBundle\Controller;
 
 use AppBundle\Entity\ImageGallery;
-use AppBundle\Entity\Image;
-use AppBundle\Form\Type\EditImageType;
 use AppBundle\Form\Type\ImageGalleryType;
-use AppBundle\Form\Type\UploadImageType;
 use AppBundle\Role\Roles;
 use Doctrine\ORM\NoResultException;
 use Liip\ImagineBundle\Exception\Imagine\Filter\NonExistingFilterException;
@@ -79,30 +76,6 @@ class ImageGalleryController extends Controller
         ));
     }
 
-    public function uploadImageAction(Request $request, ImageGallery $imageGallery)
-    {
-        $image = new Image();
-
-        $form = $this->createForm(UploadImageType::class, $image, array('validation_groups' => array('image_gallery_upload')));
-        $form->handleRequest($request);
-
-        if ($form->isSubmitted() && $form->isValid()) {
-            $image->setGallery($imageGallery);
-
-            $em = $this->getDoctrine()->getManager();
-            $em->persist($image);
-            $em->flush();
-
-            $this->get('event_dispatcher')->dispatch(ImageGalleryEvent::IMAGE_ADDED, new ImageGalleryEvent($imageGallery));
-
-            return $this->redirectToRoute('image_gallery_edit', array('id' => $imageGallery->getId()));
-        }
-        return $this->render('image_gallery/upload_image.html.twig', array(
-            'image_gallery' => $imageGallery,
-            'form' => $form->createView(),
-        ));
-    }
-
     public function deleteAction(ImageGallery $imageGallery)
     {
         $em = $this->getDoctrine()->getManager();
@@ -112,44 +85,6 @@ class ImageGalleryController extends Controller
         $this->get('event_dispatcher')->dispatch(ImageGalleryEvent::DELETED, new ImageGalleryEvent($imageGallery));
 
         return $this->redirectToRoute('image_gallery_create');
-    }
-
-    public function editImageAction(Request $request, Image $image)
-    {
-        $oldPath = $image->getPath();
-        $form = $this->createForm(EditImageType::class, $image);
-        $form->handleRequest($request);
-
-        if ($form->isSubmitted() && $form->isValid()) {
-            $image->setPath($oldPath);
-
-            $em = $this->getDoctrine()->getManager();
-            $em->persist($image);
-            $em->flush();
-
-            $imageGallery = $image->getGallery();
-            $this->get('event_dispatcher')->dispatch(ImageGalleryEvent::IMAGE_EDITED, new ImageGalleryEvent($imageGallery));
-
-            return $this->redirectToRoute('image_edit', array('id' => $image->getId()));
-        }
-
-        return $this->render('image_gallery/edit_image.html.twig', array(
-            'form' => $form->createView(),
-            'image' => $image,
-        ));
-    }
-
-    public function deleteImageAction(Image $image)
-    {
-        $imageGallery = $image->getGallery();
-
-        $em = $this->getDoctrine()->getManager();
-        $em->remove($image);
-        $em->flush();
-
-        $this->get('event_dispatcher')->dispatch(ImageGalleryEvent::IMAGE_REMOVED, new ImageGalleryEvent($imageGallery));
-
-        return $this->redirectToRoute('image_gallery_edit', array('id' => $imageGallery->getId()));
     }
 
     public function getAction(Request $request, $referenceName)
